moved docker compose to root of server repo and integrate with controllers & routes
