+++ docker-compose.yaml
@@ -1,34 +1,80 @@
 
-version: "3.8"
-
+version: "3"
 services:
-  # account:
-  #   build: "src/account/"
-  #   command: "npm run account"
-  #   ports: 
-  #     - "5000:5000" 
-  #   env_file: "src/account/.env"
-  #   dns:
-  #     - "8.8.8.8"
-  models:
-    build: "src/models/"
-    command: "echo Models Container Working"
-  routes:
-    build: "src/routes/"
-    command: "npx ts-node github.ts"
+  activemq: 
+    image: rmohr/activemq
+    container_name: activemq
+    networks:
+      - backend
+    ports: 
+      - "8161:8161"
+      - "1883:1883"
+    env_file: ".env"
+
+  # server:
+  #   build: "."
+  #   command: "npm run server"
+  #   ports:
+  #     - "4000:4000"
+  #   networks:
+  #     - backend
+  #   env_file: ".env"
+  #   depends_on:
+  #     - "activemq"
+  account:
+    container_name: "account"
+    build: "src/account/"
+    command: "npm run account"
+    ports: 
+      - "5000:5000" 
+    env_file: ".env"
+    dns:
+      - "8.8.8.8"
+    networks:
+      - backend
+  publish:
+    container_name: "publish"
+    build:
+      context: 'src/notification/publish'
+    env_file: ".env"
     ports:
-      - "4000:4000"
-      - "8080:8080"
-    env_file: "src/routes/.env"
+      - 4000:4000
+    networks:
+      - backend
     depends_on:
-      - "models"
-  client:
-    build: "client/"
-    command: "npm start"
-    ports:
-      - "3000:3000"
-    env_file: "src/routes/.env"
-# Requires server services to run before running client servers
+      - activemq
+    
+  subscribe:
+    container_name: subscribe
+    build:
+      context: 'src/notification/subscribe'
+    env_file: ".env"
+    networks:
+      - backend
     depends_on:
-      # - "account"
-      - "routes"
+      - activemq
+#   models:
+#     build: "src/models/"
+#     command: "echo Models Container Working"
+#   routes:
+#     build: "src/routes/"
+#     command: "npx ts-node github.ts"
+#     ports:
+#       - "4000:4000"
+#       - "8080:8080"
+#     env_file: "src/routes/.env"
+#     depends_on:
+#       - "models"
+#   client:
+#     build: "client/"
+#     command: "npm start"
+#     ports:
+#       - "3000:3000"
+#     env_file: "src/routes/.env"
+# # Requires server services to run before running client servers
+#     depends_on:
+#       # - "account"
+#       - "routes"
+
+networks:
+  backend:
