+++ package.json
@@ -2,27 +2,33 @@
   "name": "src",
   "version": "1.0.0",
   "description": "backend",
-  "main": "index.ts",
+  "main": "server.ts",
   "directories": {
     "test": "tests"
   },
   "scripts": {
     "test": "jest test",
-    "ts-coverage": "typescript-coverage-report"
+    "ts-coverage": "typescript-coverage-report",
+    "server": "npx ts-node src/server/server.ts"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "@types/mqtt": "^2.5.0",
     "@types/uuid": "^10.0.0",
+    "cors": "^2.8.5",
     "dotenv": "^16.4.1",
     "express": "^4.18.2",
     "jest": "^29.7.0",
+    "mongodb": "^6.7.0",
     "mongoose": "^8.1.1",
+    "mqtt": "^5.7.2",
     "react-scripts": "^5.0.1",
-    "typescript-coverage-report": "^0.7.0"
+    "typescript-coverage-report": "^0.7.0",
+    "uuid": "^10.0.0"
   },
   "devDependencies": {
+    "@types/cors": "^2.8.17",
     "@types/dotenv": "^8.2.0",
     "@types/jest": "^29.5.6",
     "@types/mongoose": "^5.11.97",
