+++ src/routes/accountRoutes.ts
@@ -4,8 +4,7 @@ import {
   getAllAccounts,
   getAccountById,
   deleteAccountById,
-  publishMessage,
-} from '../controllers/publishController';
+} from '../controllers/accountController';
 
 const router = express.Router();
 
@@ -13,6 +12,5 @@ router.post('/accounts/new', registerAccount);
 router.get('/accounts/all', getAllAccounts);
 router.get('/accounts/:id', getAccountById);
 router.delete('/accounts/:id/delete', deleteAccountById);
-router.get('/:id', publishMessage);
 
 export default router;
